restructure file test case and share some code
