+++ tests/unittests/commands/test_makemigrations.py
@@ -20,7 +20,7 @@ def with_test_output_directory(fun):
 
 
 class MakemigrationsCommandTestCase(FileTestCase):
-    target = os.path.join(this_dir, 'out/data_migrations')
+    internal_target = os.path.join(this_dir, 'out/data_migrations')
     needs_cleanup = False
 
     def tearDown(self) -> None:
