+++ tests/unittests/test_app_2/helper.py
@@ -1,9 +1,10 @@
 import os.path
+from tests.utils import ResetDirectoryMixin
 
 this_dir = os.path.dirname(__file__)
 
 
-class ResetDirectory2Context:
+class ResetDirectory2Context(ResetDirectoryMixin):
     targets = ['migrations', 'data_migrations']
     protected_files = [
         '0001_initial.py', '0004_customer_is_active.py',
@@ -12,21 +13,4 @@ class ResetDirectory2Context:
         '0008_customer_is_business.py', '0003_mmodel.py',
         '0006_address_line_split.py', '__init__.py',
     ]
-
-    def __enter__(self):
-        return True
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        for target in self.targets:
-            dir_path = os.path.join(this_dir, target)
-            try:
-                files = [
-                    os.path.join(dir_path, f)
-                    for f in os.listdir(dir_path)
-                    if f not in self.protected_files
-                    and os.path.isfile(os.path.join(dir_path, f))
-                ]
-            except FileNotFoundError:
-                continue
-            for file in files:
-                os.remove(file)
+    this_dir = this_dir
